Removed 'Main' scene because it smells; replaced with my own version.
Replaced with a "testing" scene. Thanks to Harry, we figured out why the
2D colliders weren't playing nice.
Added back stuff from 'Main' and extended. You can now control a cube.
The movement is reasonably smooth, but the player will often move
part-way into an object before being moved back. Additionally, due to
the way that the velocity interacts with the physics engine, you can
hold onto the wall indefinitely.

Regardless, the basics of movement are there, we can tweak the numbers
later.