Some small fixes
Had an issue where the framerate was affecting how fast the camera
followed the player, and how much gravity was being applied.
Additionally this fixes the problem of resetting the yVel to 0 when you
pause the game.
Also added a FPSDisplay to the game. We might want to take this out
before sending the game out.